requirements.txt
attrs==19.1.0
bokeh==1.3.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Fiona==1.8.6
Flask==1.1.1
geopandas==0.5.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
munch==2.3.2
numpy==1.17.2
packaging==19.2
pandas==0.25.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pyparsing==2.4.2
pyproj==2.3.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
tornado==6.0.3
urllib3==1.25.5
Werkzeug==0.16.0
xlrd==1.2.0
